requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
prometheus_flask_exporter